+++ src/vehicle/vehiclecomponent/StepperComt.cpp
@@ -5,7 +5,7 @@
 #include "StepperComt.h"
 
 StepperComt::StepperComt(uint16_t stepNumber, uint8_t pin1, uint8_t pin2, uint8_t pin3, uint8_t pin4,
-                         const String &acIdent, const String &stIdent, const String &rIdent, uint8_t multi,
+                         const String &acIdent, const String &stIdent, const String &rIdent, double multi,
                          long speed) : stepper(
         Stepper(stepNumber, pin1, pin2, pin3, pin4)),
                                        stepperAc(acIdent, stepper),
